Refactor: PassengerCard, create PassengerCardDetail component
