+++ packages/components/index.js
@@ -34,8 +34,8 @@ export {
 } from './src/illustrationWithInformation/IllustrationWithInformation';
 export { default as PassengerCard } from './src/passengerCard/PassengerCard';
 export {
-  default as PassengerCardDetail,
-} from './src/passengerCard/PassengerCardDetail';
+  default as PassengerCardDetailItem,
+} from './src/passengerCard/PassengerCardDetailItem';
 export { default as PassengerCards } from './src/passengerCards/PassengerCards';
 export { default as Separator } from './src/separator/Separator';
 export { default as ShareIcon } from './src/shareIcon/ShareIcon';
